Creacion del regidtro en su totalidad con sus debidas imagenes y del login guardado en base de datos

+++ app/src/main/java/com/example/lovelink/models/ImagenesUsuario.kt
@@ -1,8 +1,8 @@
 package com.example.lovelink.models
 
 data class ImagenesUsuario(
-    val id_imagen: Int? = null,
-    val id_usuario: Int,
+    val id_imagen: Long? = null,
+    val id_usuario: Long,
     val imagen1: String?,
     val imagen2: String?,
     val imagen3: String?,
